feat(v7): Deprecate and relocate `trpcMiddleware`
The `trpcMiddleware` export will move from `Handlers` to a top level export in v8. This is more or less backporting https://github.com/getsentry/sentry-javascript/pull/11374 to v7.
