+++ packages/remix/src/index.server.ts
@@ -87,6 +87,7 @@ export {
   inboundFiltersIntegration,
   linkedErrorsIntegration,
   Handlers,
+  trpcMiddleware,
   setMeasurement,
   getActiveSpan,
   startSpan,
