+++ packages/sveltekit/src/server/index.ts
@@ -81,6 +81,7 @@ export {
   inboundFiltersIntegration,
   linkedErrorsIntegration,
   Handlers,
+  trpcMiddleware,
   setMeasurement,
   getActiveSpan,
   startSpan,
